pyproject.toml
[project]
name = "link-organizer"
version = "0.2.0"
description = "A link organizer tool that extracts links from markdown and crawls/downloads content for offline storage"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crawl4ai>=0.4.0",
    "requests>=2.31.0",
    "litellm>=1.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "aiohttp>=3.9.0",
    "PyPDF2>=3.0.0",
    "Pillow>=10.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
link = "src.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = ["src/", "cli.py"]
